boreal-py/pyproject.toml
[project]
name = "boreal-python"
version = "1.0.0"
readme = "README.md"
description = "Python bindings to the boreal YARA scanner"
requires-python = ">=3.11"

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-run-parallel>=0.7.0",
    "yara-python>=4.5.1",
]
doc = [
    "mike>=2.1.3",
    "mkdocs-material>=9.6.9",
    "mkdocstrings[python]>=0.29.0",
    "ruff>=0.11.2",
]

[tool.uv]
[[tool.uv.cache-keys]]
file = "pyproject.toml"
[[tool.uv.cache-keys]]
file = "Cargo.toml"
[[tool.uv.cache-keys]]
file = "boreal.pyi"
[[tool.uv.cache-keys]]
file = "src/**/*"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
markers = [
    "global_config: tests that uses global config parameters"
]
